+++ packages/@expo/styling/next/index.ts
@@ -1,2 +1,0 @@
-import React from 'react';
-export const createElement = React.createElement;