+++ packages/@expo/styling/src/__tests__/jest.d.ts
@@ -1,3 +1,4 @@
+/// <reference types="@types/jest" />
 export {};
 declare global {
   namespace jest {