+++ packages/@expo/styling/src/index.native.ts
@@ -1,5 +1,6 @@
 import React from 'react';
 
 export { StyleSheet } from './runtime/native/stylesheet';
+export { defaultCSSInterop, makeStyled, svgCSSInterop } from './runtime/polyfill';
 
 export const createElement = React.createElement;