+++ packages/@expo/styling/src/index.ts
@@ -1,5 +1,6 @@
 import React from 'react';
 
 export { StyleSheet } from 'react-native';
+export { defaultCSSInterop, makeStyled, svgCSSInterop } from './runtime/polyfill';
 
 export const createElement = React.createElement;