+++ packages/@expo/styling/src/runtime/polyfill/mapping.ts
@@ -1,0 +1,11 @@
+import type { ComponentType } from 'react';
+
+export type InteropFunction = (
+  jsx: Function,
+  type: any,
+  props: Record<string | number, unknown>,
+  key: string,
+  experimentalFeatures?: boolean
+) => any;
+
+export const polyfillMapping = new WeakMap<ComponentType, InteropFunction>();