+++ packages/@expo/styling/src/runtime/render.ts
@@ -1,12 +1,15 @@
+import { polyfillMapping } from './polyfill/mapping';
+
 export function render(
   jsx: Function,
   type: any,
   props: Record<string | number, unknown>,
   key: string,
-  next = false
+  experimental = false
 ) {
-  if (typeof type.cssInterop === 'function' && !props.__skipCssInterop) {
-    return type.cssInterop(jsx, type, props, key, next);
+  const cssInterop = polyfillMapping.get(type);
+  if (cssInterop && !props.__skipCssInterop) {
+    return cssInterop(jsx, type, props, key, experimental);
   } else {
     return jsx(type, props, key);
   }