+++ packages/@expo/styling/src/transformer/metro-transform-worker.d.ts
@@ -1,6 +1,0 @@
-declare module 'metro-transform-worker' {
-  export type JsTransformerConfig = any;
-  export type JsTransformOptions = any;
-  export type TransformResponse = any;
-  export default any;
-}